fix : 소셜 로그인 URL 관련 cors 에러 해결
소셜 로그인 URL로 리다이렉트를 보내는 요청에서 발생했던 cors 문제를 1차 수정했습니다.
